Factor out src/Types/Core.hs
- move most types to Types.Core
- move some code to Types.Prelude
- remove some unhelpful helpers
- shuffle things around a lot
- get rid of unused Binary instances
